pyproject.toml
[tool.poetry]
name = "choirsupport"
version = "0.2.0"
description = "Exports a MIDI file to one MIDI file for each voice"
authors = ["Jan van Mansum <janvanmansum@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mido = "^1.2.9"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
create-choir-parts = "choirsupport.create_choir_parts:main"
convert-to-mp3 = "choirsupport.convert_to_mp3:main"

